e2etest: update overwrite tests to use the refactored code
